+++ src/lib/auth/Authentication.ts
@@ -1,7 +1,8 @@
 /** An interface that defines an authentication state interface. Useful if we want to migrate to other auth providers. */
-export default abstract class Authentication<UserKind> {
+export default abstract class Authentication<UserKind, ErrorKind> {
 	abstract setUser(user: UserKind | null): void;
 	abstract getUserID(): string | null;
+	abstract signIn(email: string, password: string | undefined): Promise<ErrorKind | null>;
 	abstract isAuthenticated(): boolean;
-	abstract signOut(): void;
+	abstract signOut(): Promise<ErrorKind | null>;
 }