+++ src/routes/about/+page.svelte
@@ -10,8 +10,3 @@
 	ourselves and our communities. You can learn more about who's contributing and how to contribute
 	yourself on <Link to="https://github.com/reciprocalreviews">GitHub</Link>.
 </p>
-
-<Note>
-	Why ORCID? It's the central way to manage reviewer identity. It's designed for this explicit
-	purpose and there's no reason to require another login.
-</Note>